app: Complete the implementation of the participant app
Move application bootstrap file for client and server to the proper
directory in `apps/participant` and move the SurveyRouter client route to
the client app folder.

Update the demo_survey to use the participant app both for client and
server.

The participant application is served from the public/dist/<project
shortname>/survey path.

** BREAKING **

The `app-survey` and `server` files have moved to the `apps/participant`
folder. References needs to be updated to match the new path. In this
commit, the admin routes do not work anymore.
