+++ example/demo_survey/src/app-survey.tsx
@@ -1,4 +1,4 @@
-import runClientApp from 'evolution-legacy/lib/apps/participant/app-survey';
+import runClientApp from 'evolution-legacy/lib/apps/participant/client';
 import { SurveyAppConfig } from './config/application.config';
 import { setApplicationConfiguration } from 'chaire-lib-frontend/lib/config/application.config';
 
@@ -6,26 +6,10 @@ import surveySections from './survey/sections';
 import * as widgetsConfig from './survey/widgets';
 import projectHelpers from './survey/helper';
 
-let validations = undefined;
-let parsers = undefined;
-
-// TODO Let there be an admin config that can be loaded only for the admin application, to avoid using require for the admin settings.
 setApplicationConfiguration<SurveyAppConfig>({
     sections: surveySections,
     widgets: widgetsConfig,
-    getAdminValidations: () => {
-        if (validations === undefined) {
-            validations = require(`./admin/validations.js`).default;
-        }
-        return validations;
-    },
     projectHelpers,
-    getParsers: () => {
-        if (parsers === undefined) {
-            parsers = require(`./admin/parsers.js`).default;
-        }
-        return parsers;
-    }
 });
 
 runClientApp();
