+++ example/demo_survey/src/server.ts
@@ -8,9 +8,7 @@ import path from 'path';
 
 import setupServer from 'evolution-legacy/lib/apps/participant/server';
 import { setProjectConfig } from 'evolution-backend/lib/config/projectConfig';
-import router from 'chaire-lib-backend/lib/api/admin.routes';
 import { registerTranslationDir, addTranslationNamespace } from 'chaire-lib-backend/lib/config/i18next';
-import adminRoutes from './routes/admin/admin.routes';
 
 
 const configureServer = () => {
@@ -20,8 +18,6 @@ const configureServer = () => {
         serverValidations: undefined,
         roleDefinitions: undefined
     });
-
-    adminRoutes(router);
 }
 
 setupServer(configureServer);
