+++ packages/evolution-legacy/src/apps/participant/client/index.js
@@ -12,24 +12,18 @@ import { createBrowserHistory } from 'history';
 import { Router }               from 'react-router-dom';
 
 import config from 'chaire-lib-frontend/lib/config/project.config';
-import i18n              from '../../config/survey/i18n.config';
-import SurveyRouter      from '../../routers/survey/SurveyRouter';
-import configureStore    from '../../store/survey/configureStore';
-import { login, logout } from '../../actions/shared/auth';
-import LoadingPage       from '../../components/shared/LoadingPage';
+import i18n              from '../../../config/survey/i18n.config';
+import SurveyRouter      from './SurveyRouter';
+import configureStore    from '../../../store/survey/configureStore';
+import { login, logout } from '../../../actions/shared/auth';
+import LoadingPage       from '../../../components/shared/LoadingPage';
 import { InterviewContext, interviewReducer, initialState } from 'evolution-frontend/lib/contexts/InterviewContext';
 // TODO When the project is the root of the application (instead of evolution directly importing project files), this should go in the project
 import { SurveyContext, surveyReducer } from 'evolution-frontend/lib/contexts/SurveyContext';
 import appConfig, { setApplicationConfiguration } from 'chaire-lib-frontend/lib/config/application.config';
-import '../../styles/survey/styles-survey.scss';
+import '../../../styles/survey/styles-survey.scss';
   
-// TODO This is a workaround to get the links to the user, until some more complete solution is implemented (see https://github.com/chairemobilite/transition/issues/1516)
-const pages = [
-    { path: '/admin/validation', permissions: { Interviews: ['validate'] }, title: 'admin:validationPageTitle' },
-    { path: '/admin/monitoring', permissions: { Interviews: ['manage'] }, title: 'admin:monitoringPageTitle' },
-    { path: '/admin/users', permissions: { Users: ['read', 'update'] }, title: 'admin:usersPageTitle' }
-];
-setApplicationConfiguration({ homePage: '/survey', pages });
+setApplicationConfiguration({ homePage: '/survey' });
 
 export default () => {
     document.title = config.title[i18n.language];
